+++ code/src/nuvla/connector/docker_connector.py
@@ -1,23 +1,33 @@
 # -*- coding: utf-8 -*-
 
-import re
 import json
 import logging
-import requests
+import re
 from collections import defaultdict
 from tempfile import NamedTemporaryFile
 
+import requests
+
 from .connector import Connector, ConnectorError, should_connect
 
+"""
+Service is a set of tasks. Service doesn't have a state, but tasks do.
+
+       task state
+      /    |    \
+running shutdown accepted <- DesiredState
+        /   |   \
+ shutdown failed rejected <- Status.State
+
+Single replica task spawns a new container.
+    
+"""
 
 log = logging.getLogger('docker_connector')
 
-
 bytes_per_mib = 1048576
 
-
 as_nanos = 1000000000
-
 
 tolerance = 1.1  # 10% leeway on resource requirements
 
@@ -90,13 +100,16 @@ class DockerConnector(Connector):
         working_dir = start_kwargs.get('working_dir')
         cpus = start_kwargs.get('cpus')
         memory = start_kwargs.get('memory')
-        restart_policy = start_kwargs.get('restart_policy')
         cmd = start_kwargs.get('cmd')
         args = start_kwargs.get('args')
+        restart_policy_condition = start_kwargs.get('restart_policy_condition')
+        restart_policy_delay = start_kwargs.get('restart_policy_delay')
+        restart_policy_max_attempts = start_kwargs.get('restart_policy_max_attempts')
+        restart_policy_window = start_kwargs.get('restart_policy_window')
 
         #
         # The fields that are supported by the Docker API are documented here:
-        # https://docs.docker.com/engine/api/v1.24/#39-services
+        # https://docs.docker.com/engine/api/v1.37/#39-services
         #
         service_json = tree()
 
@@ -123,8 +136,17 @@ class DockerConnector(Connector):
             service_json['TaskTemplate']['Resources']['Limits']['MemoryBytes'] = ram_bytes_hard
             service_json['TaskTemplate']['Resources']['Reservations']['MemoryBytes'] = ram_bytes_soft
 
-        if restart_policy:
-            service_json['TaskTemplate']['RestartPolicy']['Condition'] = restart_policy
+        if restart_policy_condition:
+            service_json['TaskTemplate']['RestartPolicy']['Condition'] = restart_policy_condition
+
+        if restart_policy_delay:
+            service_json['TaskTemplate']['RestartPolicy']['Delay'] = restart_policy_delay
+
+        if restart_policy_max_attempts:
+            service_json['TaskTemplate']['RestartPolicy']['MaxAttempts'] = restart_policy_max_attempts
+
+        if restart_policy_window:
+            service_json['TaskTemplate']['RestartPolicy']['Window'] = restart_policy_window
 
         if cmd:
             service_json['TaskTemplate']['ContainerSpec']['command'] = [cmd]
@@ -189,7 +211,6 @@ class DockerConnector(Connector):
         params = {'filters': convert_filters(filters) if filters else None}
         return self.docker_api.get(self._get_full_url("tasks"), params=params).json()
 
-    @should_connect
     def service_replicas(self, sname, tasks_filters=None):
         """
         Returns number of running and desired replicas of `sname` service as two-tuple
@@ -203,25 +224,37 @@ class DockerConnector(Connector):
         :param tasks_filters: dict, see `service_tasks()`->`filters` for dict spec.
         :return: (int, int)
         """
-        running = -1
-        desired = -1
+        desired = self.service_replicas_desired(sname)
+        if desired < 0:
+            return -1, -1
+        running = self.service_replicas_running(sname)
+        return running, desired
+
+    def service_replicas_desired(self, sname):
+        """
+        Returns number of desired replicas of `sname` service.
+        Returns -1 in case `sname` service is not found.
+        Returns -1 in case `sname` service is not in Replicated or Global mode.
+        -1 indicates an error.
+
+        :param sname: str, service name.
+        :param tasks_filters: dict, see `service_tasks()`->`filters` for dict spec.
+        :return: int
+        """
         services = self.list(filters={"name": sname})
         if len(services) != 1:
-            return running, desired
+            return -1
         mode = services[0]['Spec']['Mode']
         if 'Replicated' in mode:
-            desired = mode['Replicated']['Replicas']
+            return mode['Replicated']['Replicas']
         elif 'Global' in mode:
-            desired = len(self.nodes_list_active())
-        if tasks_filters:
-            tasks_filters.update({'service': sname})
+            return len(self.nodes_list_active())
         else:
-            tasks_filters = {'service': sname}
-        tasks = self.service_tasks(filters=tasks_filters)
-        return len(tasks), desired
+            return -1
 
     def service_replicas_running(self, sname):
-        return self.service_replicas(sname, tasks_filters={'desired-state': 'running'})
+        return self.service_tasks(filters={'service': sname,
+                                           'desired-state': 'running'})
 
     @should_connect
     def nodes_list(self, availability=None):